test/quantum/aws/requirements.txt
boto3
amazon-braket-sdk
